# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ClaimResponse:
    properties:
      answer_input:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      item_id:
        type: string
      owner_id:
        type: string
      status:
        type: string
    type: object
  dto.ContactRequest:
    properties:
      platform:
        example: WHATSAPP
        type: string
      value:
        example: "08123456789"
        type: string
    required:
    - platform
    - value
    type: object
  dto.ContactResponse:
    properties:
      platform:
        type: string
      value:
        type: string
    type: object
...
  dto.CreateClaimRequest:
    properties:
      answer_input:
        example: Blue wallet with university ID
        type: string
      image_url:
        example: http://example.com/proof.jpg
        type: string
    required:
    - answer_input
    type: object
  dto.CreateFoundItemRequest:
    properties:
      category_id:
        example: 1bd43cf7-fc4f-4968-bd4f-c45699b03c18
        type: string
      cod:
        example: false
        type: boolean
      contacts:
        items:
          $ref: '#/definitions/dto.ContactRequest'
        type: array
      date_found:
        description: Format YYYY-MM-DD or RFC3339
        example: "2023-10-27"
        type: string
      image_url:
        example: http://example.com/image.jpg
        type: string
      location_id:
        example: e9464495-bfe5-4ed0-8ea4-a2d69afa0b39
        type: string
      return_method:
        enum:
        - BRING_BY_FINDER
        - HANDED_TO_SECURITY
        example: BRING_BY_FINDER
        type: string
      show_phone:
        example: false
        type: boolean
      title:
        example: Blue Wallet
        type: string
      verifications:
        items:
          $ref: '#/definitions/dto.VerificationRequest'
        type: array
    required:
    - category_id
    - date_found
    - location_id
    - return_method
    - title
    - verifications
    type: object
  dto.CreateLostItemRequest:
    properties:
      category_id:
        example: 1bd43cf7-fc4f-4968-bd4f-c45699b03c18
        type: string
      contacts:
        items:
          $ref: '#/definitions/dto.ContactRequest'
        type: array
      date_lost:
        description: Format YYYY-MM-DD
        example: "2023-10-26"
        type: string
      description:
        example: Black case with a sticker
        type: string
      image_url:
        example: http://example.com/iphone.jpg
        type: string
      location_last_seen:
        example: Canteen
        type: string
      offer_reward:
        example: true
        type: boolean
      show_phone:
        example: false
        type: boolean
      title:
        example: iPhone 13
        type: string
      urgency:
        enum:
        - NORMAL
        - HIGH
        - CRITICAL
        example: HIGH
        type: string
    required:
    - category_id
    - date_lost
    - location_last_seen
    - title
    type: object
  dto.DecideClaimRequest:
    properties:
      status:
        enum:
        - APPROVED
        - REJECTED
        type: string
    required:
    - status
    type: object
  dto.ItemResponse:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      contacts:
        description: For both item types
        items:
          $ref: '#/definitions/dto.ContactResponse'
        type: array
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      location_id:
        type: string
      location_name:
        type: string
      offer_reward:
        description: For lost items
        type: boolean
      show_phone:
        description: For both item types
        type: boolean
      status:
        type: string
      title:
        type: string
      urgency:
        description: For lost items
        type: string
      verifications:
        description: For found items
        items:
          $ref: '#/definitions/dto.VerificationResponse'
        type: array
    type: object
...
  dto.VerificationRequest:
    properties:
      answer:
        example: Blue
        type: string
      question:
        example: What is the color?
        type: string
    required:
    - answer
    - question
    type: object
  dto.VerificationResponse:
    properties:
      question:
        type: string
    type: object
...
  models.CampusLocation:
    properties:
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  models.Claim:
    properties:
      answer_input:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        description: Proof image for claim
        type: string
      item:
        $ref: '#/definitions/models.Item'
      item_id:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      status:
        $ref: '#/definitions/models.ClaimStatus'
      updated_at:
        type: string
    type: object
  models.ClaimStatus:
    enum:
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - ClaimStatusPending
    - ClaimStatusApproved
    - ClaimStatusRejected
...
  models.Item:
    properties:
      category:
        $ref: '#/definitions/models.ItemCategory'
      category_id:
        type: string
      cod:
        type: boolean
      contacts:
        items:
          $ref: '#/definitions/models.ItemContact'
        type: array
      created_at:
        type: string
      date_found:
        type: string
      date_lost:
        type: string
      finder:
        $ref: '#/definitions/models.User'
      finder_id:
        description: Nullable for Lost items
        type: string
      id:
        type: string
      image_url:
        type: string
      location:
        $ref: '#/definitions/models.CampusLocation'
      location_description:
        description: For Lost items (free text)
        type: string
      location_id:
        description: Nullable for Lost items
        type: string
      offer_reward:
        type: boolean
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        description: Nullable for Found items
        type: string
      return_method:
        $ref: '#/definitions/models.ReturnMethod'
      show_phone:
        type: boolean
      status:
        $ref: '#/definitions/models.ItemStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/models.ItemType'
      updated_at:
        type: string
      urgency:
        $ref: '#/definitions/models.ItemUrgency'
      verifications:
        items:
          $ref: '#/definitions/models.ItemVerification'
        type: array
    type: object
  models.ItemCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.ItemContact:
    properties:
      created_at:
        type: string
      id:
        type: string
      item_id:
        type: string
      platform:
        $ref: '#/definitions/models.PlatformType'
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.ItemStatus:
    enum:
    - OPEN
    - CLAIMED
    - RESOLVED
    type: string
    x-enum-varnames:
    - ItemStatusOpen
    - ItemStatusClaimed
    - ItemStatusResolved
  models.ItemType:
    enum:
    - LOST
    - FOUND
    type: string
    x-enum-varnames:
    - ItemTypeLost
    - ItemTypeFound
  models.ItemUrgency:
    enum:
    - NORMAL
    - HIGH
    - CRITICAL
    type: string
    x-enum-varnames:
    - UrgencyNormal
    - UrgencyHigh
    - UrgencyCritical
  models.ItemVerification:
    properties:
      created_at:
        type: string
      id:
        type: string
      item_id:
        type: string
      question:
        type: string
      updated_at:
        type: string
    type: object
...
  models.PlatformType:
    enum:
    - INSTAGRAM
    - TELEGRAM
    - LINE
    - TWITTER
    - EMAIL
    - WHATSAPP
    - OTHER
    type: string
    x-enum-varnames:
    - PlatformInstagram
    - PlatformTelegram
    - PlatformLine
    - PlatformTwitter
    - PlatformEmail
    - PlatformWhatsapp
    - PlatformOther
  models.ReturnMethod:
    enum:
    - BRING_BY_FINDER
    - HANDED_TO_SECURITY
    type: string
    x-enum-varnames:
    - ReturnMethodBringByFinder
    - ReturnMethodHandedToSecurity
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      faculty:
        description: Nullable, null for Staff/Dosen
        type: string
      id:
        type: string
      identity_number:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  models.UserRole:
    enum:
    - PUBLIK
    - MAHASISWA
    - STAFF_DOSEN
    - ADMIN
    - SECURITY
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleStudent
    - RoleStaff
    - RoleAdmin
    - RoleSecurity
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Campus Lost & Found System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Campus Lost & Found API
  version: "1.0"
paths:
...
  /claims/{id}/decide:
    put:
      consumes:
      - application/json
      description: Decide on a claim (Finder only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Decide Claim Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DecideClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve or Reject a claim
      tags:
      - claims
...
  /items:
    get:
      consumes:
      - application/json
      description: Get a list of all items (Lost & Found) with optional filters
      parameters:
      - description: Filter by status (e.g., OPEN, CLAIMED)
        in: query
        name: status
        type: string
      - description: Filter by type (LOST, FOUND)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all items
      tags:
      - items
  /items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an item (Finder or Owner only)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific item
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get item by ID
      tags:
      - items
  /items/{id}/claim:
    post:
      consumes:
      - application/json
      description: Claim a found item
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Create Claim Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClaimResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a claim for an item
      tags:
      - items
  /items/{id}/claims:
    get:
      consumes:
      - application/json
      description: Get all claims for an item (Finder only)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Claim'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get claims for an item
      tags:
      - items
  /items/found:
    post:
      consumes:
      - application/json
      description: Report an item found without QR code
      parameters:
      - description: Create Found Item Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFoundItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Report a found item (Finder First)
      tags:
      - items
  /items/lost:
    post:
      consumes:
      - application/json
      description: Report a lost item without QR code
      parameters:
      - description: Create Lost Item Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLostItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Report a lost item (Ad-Hoc)
      tags:
      - items
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"